refactor(search): consolidate AI search retriever API and internals
- unify config under ai.retriever (Retriever.local/cloudflare/from)
- merge rag.ts into retriever.ts; rename rag-source -> retriever-source
- single POST /api/search endpoint (was /api/search/{rag,retrieve})
- drop 'rag' terminology in favor of AI search across keys, env var,
  virtual module, and cache dir

Amp-Thread-ID: https://ampcode.com/threads/T-019f16a6-5bc6-711c-8e7a-ca829dc99cdd
