+++ src/globals.d.ts
@@ -28,8 +28,8 @@ declare module 'virtual:vocs/search-index' {
   export function getSearchIndex(): Promise<string>
 }
 
-declare module 'virtual:vocs/rag-index' {
-  export function getRagIndex(): Promise<string>
+declare module 'virtual:vocs/ai-search-index' {
+  export function getAiSearchIndex(): Promise<string>
 }
 
 declare module 'virtual:vocs/group-icons.css' {}
