+++ src/internal/vite-plugins.ts
@@ -14,7 +14,7 @@ import * as Llms from './llms.js'
 import * as Mdx from './mdx.js'
 import type * as OpenApi from './openapi/index.js'
 import * as OpenApiRegistry from './openapi/registry.js'
-import * as Rag from './rag.js'
+import * as Retriever from './retriever.js'
 import { SearchDocuments, SearchIndex } from './search.js'
 import * as ShikiTransformers from './shiki-transformers.js'
 import * as TaskRunner from './task-runner.js'
@@ -757,32 +757,33 @@ export function search(config: Config.Config): PluginOption {
 }
 
 /**
- * Vite plugin for RAG (semantic) search.
+ * Vite plugin for AI (semantic) search.
  *
- * When `search.rag` is enabled it builds the static vector index at build/dev
+ * When a local `ai.retriever` provider is enabled it builds the static vector index at build/dev
  * start (warming the embedding cache), and — for `runtime: 'client'` or
  * `vectorStore.expose` — emits a public browser artifact and serves it via the
- * `virtual:vocs/rag-index` module. Server-runtime search is handled at runtime
- * by the `/api/search/rag` endpoint (which builds an in-memory index), so this
+ * `virtual:vocs/ai-search-index` module. Server-runtime search is handled at runtime
+ * by the `/api/search` endpoint (which builds an in-memory index), so this
  * plugin's main job there is cache warming.
  *
  * @param config - Vocs configuration.
  * @returns Plugin.
  */
-export function ragSearch(config: Config.Config): PluginOption {
-  const rag = Rag.fromConfig(config)
-  if (!rag) return []
+export function aiSearch(config: Config.Config): PluginOption {
+  const ai = Retriever.fromConfig(config)
+  if (!ai) return []
 
-  const virtualModuleId = 'virtual:vocs/rag-index'
+  const virtualModuleId = 'virtual:vocs/ai-search-index'
   const resolvedVirtualModuleId = `\0${virtualModuleId}`
   const { rootDir, basePath } = config
-  const exposePublic = rag.runtime === 'client' || config._rag?.vectorStore.expose === true
+  const exposePublic =
+    ai.runtime === 'client' || config._localRetriever?.vectorStore.expose === true
 
   let mode: 'development' | 'production' = 'development'
-  let manifestPromise: Promise<Rag.IndexManifest> | undefined
+  let manifestPromise: Promise<Retriever.IndexManifest> | undefined
   let publicHash: string | undefined
 
-  function toPublicManifest(manifest: Rag.IndexManifest): Rag.IndexManifest {
+  function toPublicManifest(manifest: Retriever.IndexManifest): Retriever.IndexManifest {
     // Strip full chunk text — the browser artifact only needs snippets.
     return {
       ...manifest,
@@ -790,25 +791,25 @@ export function ragSearch(config: Config.Config): PluginOption {
     }
   }
 
-  function buildManifest(): Promise<Rag.IndexManifest> {
+  function buildManifest(): Promise<Retriever.IndexManifest> {
     if (!manifestPromise) {
-      logger.info('Building RAG index...', { timestamp: true })
-      manifestPromise = Rag.buildIndex(config)
+      logger.info('Building AI search index...', { timestamp: true })
+      manifestPromise = Retriever.buildIndex(config)
         .then(async (manifest) => {
           if (exposePublic) {
             const json = JSON.stringify(toPublicManifest(manifest))
             publicHash = crypto.createHash('md5').update(json).digest('hex').slice(0, 12)
           }
-          await Rag.saveIndex(config, manifest)
+          await Retriever.saveIndex(config, manifest)
           logger.info(
-            `RAG index built (${manifest.vectorStore.count} chunks, ${manifest.vectorStore.format}).`,
+            `AI search index built (${manifest.vectorStore.count} chunks, ${manifest.vectorStore.format}).`,
             { timestamp: true },
           )
           return manifest
         })
         .catch((error) => {
           logger.error(
-            `Failed to build RAG index: ${error instanceof Error ? error.message : String(error)}`,
+            `Failed to build AI search index: ${error instanceof Error ? error.message : String(error)}`,
           )
           throw error
         })
@@ -826,7 +827,7 @@ export function ragSearch(config: Config.Config): PluginOption {
     if (process.env['VOCS_SKIP_SEARCH_INDEX'] !== 'true') return false
     if (exposePublic) {
       logger.warn(
-        'VOCS_SKIP_SEARCH_INDEX ignored: client-runtime RAG (exposed vector store) needs the index built at build time. Building it anyway.',
+        'VOCS_SKIP_SEARCH_INDEX ignored: client-runtime AI search (exposed vector store) needs the index built at build time. Building it anyway.',
         { timestamp: true },
       )
       return false
@@ -835,7 +836,7 @@ export function ragSearch(config: Config.Config): PluginOption {
   }
 
   return {
-    name: 'vocs:rag-search',
+    name: 'vocs:ai-search',
     configResolved(resolvedConfig) {
       mode = resolvedConfig.command === 'build' ? 'production' : 'development'
     },
@@ -847,7 +848,7 @@ export function ragSearch(config: Config.Config): PluginOption {
       // client-runtime search builds it lazily when the virtual module loads.
       if (mode !== 'production') return
       if (skipSeeding()) {
-        logger.info('Skipping RAG index build (VOCS_SKIP_SEARCH_INDEX).', { timestamp: true })
+        logger.info('Skipping AI search index build (VOCS_SKIP_SEARCH_INDEX).', { timestamp: true })
         return
       }
       void buildManifest()
@@ -859,14 +860,14 @@ export function ragSearch(config: Config.Config): PluginOption {
     async load(id) {
       if (id !== resolvedVirtualModuleId) return
       if (!exposePublic)
-        return `export const getRagIndex = async () => { throw new Error('[vocs] RAG client index is not exposed (server runtime).') }`
+        return `export const getAiSearchIndex = async () => { throw new Error('[vocs] AI search client index is not exposed (server runtime).') }`
 
       const manifest = toPublicManifest(await buildManifest())
       if (mode === 'development')
-        return `export const getRagIndex = async () => ${JSON.stringify(JSON.stringify(manifest))}`
+        return `export const getAiSearchIndex = async () => ${JSON.stringify(JSON.stringify(manifest))}`
 
       const base = basePath.endsWith('/') ? basePath : `${basePath}/`
-      return `export const getRagIndex = async () => await (await fetch("${base}assets/rag-index-${publicHash}.json")).text()`
+      return `export const getAiSearchIndex = async () => await (await fetch("${base}assets/ai-search-index-${publicHash}.json")).text()`
     },
     async writeBundle(options) {
       if (!exposePublic) return
@@ -875,13 +876,13 @@ export function ragSearch(config: Config.Config): PluginOption {
       const outDir = options.dir ?? path.resolve(rootDir, config.outDir)
       const dir = path.resolve(outDir, 'assets')
       await fs.mkdir(dir, { recursive: true })
-      await fs.writeFile(path.join(dir, `rag-index-${publicHash}.json`), json, 'utf-8')
+      await fs.writeFile(path.join(dir, `ai-search-index-${publicHash}.json`), json, 'utf-8')
 
       const bytes = Buffer.byteLength(json)
-      const max = config._rag?.vectorStore.maxClientBytes
+      const max = config._localRetriever?.vectorStore.maxClientBytes
       if (max && bytes > max)
         logger.warn(
-          `RAG client index is ${(bytes / 1e6).toFixed(1)}MB, exceeds maxClientBytes (${(max / 1e6).toFixed(1)}MB). Consider int8 format or fewer chunks.`,
+          `AI search client index is ${(bytes / 1e6).toFixed(1)}MB, exceeds maxClientBytes (${(max / 1e6).toFixed(1)}MB). Consider int8 format or fewer chunks.`,
           { timestamp: true },
         )
     },
@@ -923,7 +924,7 @@ export function virtualConfig(config: Config.Config): PluginOption {
           if (mod) server.moduleGraph.invalidateModule(mod)
           Config.setGlobal(newConfig)
           // Send the serialized (public) config only — `serializeFunctions`
-          // strips `_`-prefixed fields (e.g. `_feedback`, `_rag`) so secrets in
+          // strips `_`-prefixed fields (e.g. `_feedback`, `_localRetriever`) so secrets in
           // server-only adapters never reach the browser over the HMR channel.
           server.ws.send({
             type: 'custom',
