+++ src/vite.ts
@@ -16,6 +16,7 @@ export async function vocs(): Promise<PluginOption[]> {
   const config = await Config.resolve()
 
   return [
+    Plugins.aiSearch(config),
     Plugins.arraybuffer(),
     Plugins.deps(),
     Plugins.groupIcons(config),
@@ -31,7 +32,6 @@ export async function vocs(): Promise<PluginOption[]> {
     Plugins.llms(config),
     Plugins.mdx(config),
     Plugins.openapi(config),
-    Plugins.ragSearch(config),
     Plugins.routeWatcher(config),
     Plugins.search(config),
     Plugins.sitemap(config),
